non comp err

+++ p_1/src/util.cpp
@@ -8,7 +8,7 @@ int write_config(std::string filename, double size_enclosure,
 
 	ofs << std::fixed << std::setprecision(3) << size_enclosure << ' '
 		<< time_step << ' ' << o_aos.x.size() << '\n';
-	for (int i = 0; i < o_aos.x.size(); ++i)
+	for (unsigned long i = 0; i < o_aos.x.size(); ++i)
 		ofs << std::fixed << std::setprecision(3) << o_aos.x[i] << ' '
 			<< o_aos.y[i] << ' '  << o_aos.z[i] << ' '
 			<< o_aos.vx[i] << ' ' << o_aos.vy[i] << ' '
